Fixed issue with title for notes that start with newlines, and added a default “Note” title for notes that are saved as empty.
Added comments throughout.

Added NoteDetail delegate to properly use MVC to save the note and have the list refresh when on the iPad.

Added an NSTimer to update the modification date on notes being edited when the time changes.

Added UIAlertViews to warn against canceling a note with content and saving a note with no content.

Cleaned up iPad wonkiness.